Moved config.proto from core/framework to core/protobuf. Change: 116396958
