+++ tensorflow_serving/servables/tensorflow/session_bundle_config.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 
 import "google/protobuf/wrappers.proto";
-import "tensorflow/core/framework/config.proto";
+import "tensorflow/core/protobuf/config.proto";
 
 package tensorflow.serving;
